+++ examples/restful/market/tradingDayTicker.ts
@@ -7,9 +7,10 @@ const baseURL = process.env.BINANCE_BASE_URL || '';
 const client = new Spot('', '', { baseURL: baseURL });
 
 const options: RestMarketTypes.tradingDayTickerOptions = {
+    symbol: 'BNBUSDT',
     type: 'MINI',
 };
 
-client.tradingDayTicker('BNBUSDT', options).then((res: RestMarketTypes.tradingDayTickerResponse | RestMarketTypes.tradingDayTickerResponse[]) => {
+client.tradingDayTicker(options).then((res: RestMarketTypes.tradingDayTickerResponse | RestMarketTypes.tradingDayTickerResponse[]) => {
     console.log(res);
 }).catch(err => { console.log(err); });